Add massacre AI
A kind of iterative-BFS. Doesn't do great for scenarios with distances
between white and black pieces. Add manhatten distance to heuristic?
Not even close to AlphaBeta-pruner anymore either, that doesn't work for
massacre as there is no minimizing agent.
Also remove a bunch of stuff specific to Part B that is unnecessary for
Part A.